requirements.txt
streamlit
haystack-ai
qdrant-haystack
cohere-haystack
openpyxl
pandas
python-dotenv
